requirements.txt
numpy==1.21.1
torch==1.9.0
Pillow==8.3.1
torchvision==0.10.0
tensorboard==2.5.0
tqdm==4.62.0